ASoC: codecs: tas5825m: change the initialisation sequence to avoid resetting the coefficient when the i2s clock is enabled
The TAS5825M will by default reset the DSP registers the first time
the i2s clock is enabled. This would cause all registers set through
the mixer interface to be reset to their defaults.

Setting the lower bit of the 0x46 control register (book 0, page 0)
to 0 seems to disable this behaviour.